[s3] Properly exclude empty folders during 'listdir'
Users can create empty folders on S3, by creating an object which name
ends with a forward slash, for example `folder/`. Using the `listdir`
operation on an empty folder returns `[], ["."]`. This is unexpected,
there are no file named `.` in the folder.

The reason for its presence is AWS S3 returning the following content
when listing the content of the directory:
```
{
    'Contents': [
        {... 'Key': 'folder/', ...}
    ],
   ...
}
```
Then, `posixpath.relpath("folder/", "folder/")` returns ".".